+++ src/main/java/models/RoomModel.java
@@ -15,7 +15,7 @@ public class RoomModel implements Serializable {
     @Inject
     private AdministrationService administrationService;
     
-    private int ID;
+    private Long ID;
     private ERoomDescriptor descriptor;
     private String number;
     private int capacity;
@@ -35,11 +35,11 @@ public class RoomModel implements Serializable {
         this.administrationService = administrationService;
     }
 
-    public int getID() {
+    public Long getID() {
         return ID;
     }
 
-    public void setID(int ID) {
+    public void setID(Long ID) {
         this.ID = ID;
     }
